+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Added 'ubuntu-24.04-arm' runner to compatibility workdlow
+
 ## [8.6.6] - 2024-11-17
 
 ### Changed